add para page 59 page63; change format file;

+++ FUNC_HANDLER/modbus.h
@@ -46,25 +46,25 @@
 #define MODBUS_WRITE_REG_ADDRESS 0x5001  //寄存器写地址
 
 #define MODBUS_SEND_TIME_PERIOD 500  // 0.5秒钟
-#define RX_CHECK_TICK_TIME 20  //超过20个字节的时间没有收到新的字符，认为接收结束，10*10*1000/9600,
+#define RX_CHECK_TICK_TIME      20   //超过20个字节的时间没有收到新的字符，认为接收结束，10*10*1000/9600,
 
-#define SLAVE_ID 0x02  //设备端地址
-#define BUS_EN 0X5A
-#define BUS_DIS 0X00
-#define BUS_FUN_03H 0x03  //读寄存器地址命令
-#define BUS_FUN_06H 0x06  //写寄存器地址命令
-#define BUS_FUN_10H 0x10
-#define CMD_NUMBER (sizeof(modbusCmdlib) / sizeof(modbosCmd_t))
+#define SLAVE_ID     0x02  //设备端地址
+#define BUS_EN       0X5A
+#define BUS_DIS      0X00
+#define BUS_FUN_03H  0x03  //读寄存器地址命令
+#define BUS_FUN_06H  0x06  //写寄存器地址命令
+#define BUS_FUN_10H  0x10
+#define CMD_NUMBER   (sizeof(modbusCmdlib) / sizeof(modbosCmd_t))
 #define CHECK_NUMBER (sizeof(dataCheckLib) / sizeof(dataCheckCmd_t))
 
 typedef struct modbosCmd
 {
-    u8 modbusEn;
-    u8 slaveID;
-    u8 funCode;
-    u8 length;
-    u8 timeout;
-    u8 mode;
+    u8  modbusEn;
+    u8  slaveID;
+    u8  funCode;
+    u8  length;
+    u8  timeout;
+    u8  mode;
     u16 modePara;
     u16 VPAddr;
     u16 slaveAddr;
@@ -72,8 +72,8 @@ typedef struct modbosCmd
 } modbosCmd_t;
 typedef struct dataCheckCmd
 {
-    u8 Enable;
-    u8 page;
+    u8  Enable;
+    u8  page;
     u16 dataAddr;
     u16 backAddr;
     u16 flagAddr;
@@ -96,7 +96,7 @@ extern _TKS_FLAGA_type modbusFlag;
 
 void Modbus_UART_Init(void);
 void Modbus_Process_Task(void);
-void getCmd(u8 *index);
+void getCmd(u16 *index);
 void checkChange(void);
 void forcedOutputHnadle(void);
